+++ agora/spec.py
@@ -1,4 +1,3 @@
-import copy
   # djoser_verifier: agora.utils.djoser_verifier
   # userid_extractor: agora.utils.userid_extractor
 
@@ -11,6 +10,12 @@ SERVICE_FIELDS_COMMON = {
         '.flag.filterable': {},
         '.flag.searchable': {},
         '.flag.orderable': {}},
+    'url': {
+        '.field.string': {},
+        '.flag.nullable.default': {}},
+    'endpoint': {
+        '.field.string': {},
+        '.flag.nullable.default': {}},
     'short_description': {
         '.field.string': {},
         '.flag.searchable': {},
@@ -19,6 +24,9 @@ SERVICE_FIELDS_COMMON = {
         '.field.string': {},
         '.flag.nullable.default': {}},
     'description_internal': {
+        '.field.string': {},
+        '.flag.nullable.default': {}},
+    'tagline': {
         '.field.string': {},
         '.flag.nullable.default': {}},
     'service_type': {
@@ -32,7 +40,25 @@ SERVICE_FIELDS_COMMON = {
     'funders_for_service': {
         '.field.string': {},
         '.flag.nullable.default': {}},
-    'value_to_customer': {
+    'user_value': {
+        '.field.string': {},
+        '.flag.nullable.default': {}},
+    'target_customers': {
+        '.field.string': {},
+        '.flag.nullable.default': {}},
+    'target_users': {
+        '.field.string': {},
+        '.flag.nullable.default': {}},
+    'screenshots_videos': {
+        '.field.string': {},
+        '.flag.nullable.default': {}},
+    'languages': {
+        '.field.string': {},
+        '.flag.nullable.default': {}},
+    'standards': {
+        '.field.string': {},
+        '.flag.nullable.default': {}},
+    'certifications': {
         '.field.string': {},
         '.flag.nullable.default': {}},
     'risks': {
@@ -47,27 +73,110 @@ SERVICE_FIELDS_COMMON = {
     'internal': {
         '.flag.filterable': {},
         '.field.boolean': {}},
-    'service_area_ext': {
+    'tags': {
         '.field.string': {},
-        '.flag.nowrite': {},
+        '.flag.searchable': {},
+        '.flag.nullable.default': {}},
+    'scientific_fields': {
+        '.field.string': {},
+        '.flag.searchable': {},
+        '.flag.nullable.default': {}},
+    'owner_name': {
+        '.field.string': {},
         '.flag.orderable': {},
-        'source': 'service_area.name'},
+        '.flag.nullable.default': {},
+        '.flag.filterable': {}},
+    'owner_contact': {
+        '.field.string': {},
+        '.flag.orderable': {},
+        '.flag.nullable.default': {},
+        '.flag.filterable': {}},
+    'support_name': {
+        '.field.string': {},
+        '.flag.orderable': {},
+        '.flag.nullable.default': {},
+        '.flag.filterable': {}},
+    'support_contact': {
+        '.field.string': {},
+        '.flag.orderable': {},
+        '.flag.nullable.default': {},
+        '.flag.filterable': {}},
+    'security_name': {
+        '.field.string': {},
+        '.flag.orderable': {},
+        '.flag.nullable.default': {},
+        '.flag.filterable': {}},
+    'security_contact': {
+        '.field.string': {},
+        '.flag.orderable': {},
+        '.flag.nullable.default': {},
+        '.flag.filterable': {}},
+    'helpdesk': {
+        '.field.string': {},
+        '.flag.orderable': {},
+        '.flag.nullable.default': {},
+        '.flag.filterable': {}},
+    'order': {
+        '.field.string': {},
+        '.flag.orderable': {},
+        '.flag.nullable.default': {},
+        '.flag.filterable': {}},
+    'order_type': {
+        '.field.string': {},
+        '.flag.nullable.default': {}},
+    'changelog': {
+        '.field.string': {},
+        '.flag.nullable.default': {}},
+    'last_update': {
+        '.field.string': {},
+        '.flag.orderable': {},
+        '.flag.nullable.default': {}},
+    'service_categories_names': {
+        '.field.string': {},
+        '.flag.nowrite': {}},
+    'service_categories': {
+        '.field.collection.django': {},
+        '.flag.nullable.default': {},
+        ':filter_compat': True,
+        'flat': True,
+        'id_field': 'service_category',
+        'model': 'service.models.Service.service_categories.through',
+        'source': 'service_categories',
+        'bound': 'service',
+        'fields': {
+            'service_category': {'.field.ref': {},
+                                 'source': 'servicecategory_id',
+                                 'to': 'api/v2/service-categories'},
+        },
+    },
     'service_trl_ext': {
         '.field.string': {},
         '.flag.nowrite': {},
         '.flag.orderable': {},
         'source': 'service_trl.value'},
+    'providers_names': {
+        '.field.string': {},
+        'source': 'organisations_names',
+        '.flag.nowrite': {}},
+    'providers': {
+        '.field.collection.django': {},
+        '.flag.nullable.default': {},
+        ':filter_compat': True,
+        'flat': True,
+        'id_field': 'organisation',
+        'model': 'service.models.Service.organisations.through',
+        'source': 'organisations',
+        'bound': 'service',
+        'fields': {
+            'organisation': {'.field.ref': {},
+                            'source': 'organisation_id',
+                            'to': 'api/v2/providers'},
+        }
+    }
 }
 
 
 SERVICE_FIELDS_INT = {
-    'service_area': {
-        '.field.ref': {},
-        'source': 'service_area_id',
-        'to': '/api/v2/service-areas',
-        '.flag.nullable.default': {},
-        '.flag.orderable': {},
-        '.flag.filterable': {}},
     'service_trl': {
         '.field.ref': {},
         'source': 'service_trl_id',
@@ -306,8 +415,10 @@ SERVICE_STATUS = {
             '.field.string': {},
             '.flag.orderable': {},
             '.flag.filterable': {}},
-        'order': {
-            '.field.integer': {}},
+        'description': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {}},
         },
     'actions': {
         '.action-template.django.list': {},
@@ -390,9 +501,9 @@ USER_ROLES = {
     },
 }
 
-SERVICE_AREAS = {
+SERVICE_CATEGORIES = {
     '.collection.django': {},
-    'model': 'service.models.ServiceArea',
+    'model': 'service.models.ServiceCategory',
     'fields': {
         'id': {
             '.field.uuid': {},
@@ -500,6 +611,22 @@ CUSTOM_USERS = {
             '.flag.orderable': {},
             '.flag.filterable': {},
             '.field.string': {}},
+        'providers': {
+            '.field.collection.django': {},
+            ':filter_compat': True,
+            '.flag.nullable.default': {},
+            'flat': True,
+            'id_field': 'organisation',
+            'model': 'accounts.models.User.organisations.through',
+            'source': 'organisations',
+            'bound': 'user',
+            'fields': {
+                'organisation': {'.field.ref': {},
+                                'source': 'organisation_id',
+                                'to': 'api/v2/providers'},
+            }
+        }
+
     },
     'actions': {
         '.action-template.django.list': {},
@@ -812,11 +939,11 @@ SERVICE_TYPES_FIELDS = {
         '.flag.nowrite': {},
         '.flag.filterable': {},
         'source': 'service_id.internal'},
-    'service_area': {
+    'service_category': {
         '.field.string': {},
         '.flag.nowrite': {},
         '.flag.filterable': {},
-        'source': 'service_id.service_area.name'},
+        'source': 'service_id.service_category.name'},
     'service_version': {
         '.field.string': {},
         '.flag.nowrite': {},
@@ -883,9 +1010,9 @@ SERVICE_VERSIONS = {
         'features_future': {
             '.field.string': {},
             '.flag.nullable.default': {}},
-        'usage_policy_has': {
+        'terms_of_use_has': {
             '.field.boolean': {}},
-        'usage_policy_url': {
+        'terms_of_use_url': {
             '.field.string': {},
             '.flag.nullable.default': {}},
         'privacy_policy_has': {
@@ -895,14 +1022,16 @@ SERVICE_VERSIONS = {
             '.flag.nullable.default': {}},
         'user_documentation_has': {
             '.field.boolean': {}},
-        'user_documentation_url': {
+        'user_manual': {
             '.field.string': {},
+            'source': 'user_documentation_url',
             #'.flag.blankable': {},
             '.flag.nullable.default': {}},
         'operations_documentation_has': {
             '.field.boolean': {}},
-        'operations_documentation_url': {
+        'admin_manual': {
             '.field.string': {},
+            'source': 'operations_documentation_url',
             '.flag.nullable.default': {}},
         'monitoring_has': {
             '.field.boolean': {}},
@@ -948,6 +1077,22 @@ SERVICE_VERSIONS = {
             '.flag.orderable': {},
             '.field.boolean': {},
             'default': False},
+        'sla_url': {
+            '.field.string': {},
+            '.flag.nullable.default': {}},
+        'training_information': {
+            '.field.string': {},
+            '.flag.nullable.default': {}},
+        'maintenance_url': {
+            '.field.string': {},
+            '.flag.nullable.default': {}},
+        'service_trl': {
+            '.field.ref': {},
+            'source': 'service_trl_id',
+            'to': '/api/v2/service-trls',
+            '.flag.orderable': {},
+            '.flag.filterable': {},
+            '.flag.nullable.default': {}},
         # extended fields
         'id_service_ext': {
             '.field.string': {},
@@ -959,7 +1104,22 @@ SERVICE_VERSIONS = {
             '.flag.orderable': {},
             '.flag.nowrite': {},
             'source': 'status.value'},
-        },
+        'access_policies': {
+            '.field.collection.django': {},
+            ':filter_compat': True,
+            '.flag.nullable.default': {},
+            'flat': True,
+            'id_field': 'access_policy',
+            'model': 'service.models.ServiceDetails.access_policies.through',
+            'source': 'access_policies',
+            'bound': 'servicedetails',
+            'fields': {
+                'access_policy': {'.field.ref': {},
+                                'source': 'accesspolicy_id',
+                                'to': 'api/v2/access-policies'},
+            }
+        }
+    },
     'actions': {
         '.action-template.django.list': {},
         '.action-template.django.retrieve': {},
@@ -1070,6 +1230,119 @@ SERVICE_ADMINS = {
     },
 }
 
+ORGANISATIONS = {
+    '.collection.django': {},
+    'model': 'accounts.models.Organisation',
+    'fields': {
+        'id': {
+            '.field.uuid': {},
+            '.flag.nowrite': {}},
+        'name': {
+            '.field.string': {},
+            '.flag.orderable': {},
+            '.flag.searchable': {}},
+        'description': {
+            '.field.string': {},
+            '.flag.nullable.default': {}},
+         'contact': {
+            '.field.string': {},
+            '.flag.orderable': {},
+            '.flag.searchable': {},
+            '.flag.nullable.default': {}},
+        'logo': {
+            '.field.file': {},
+            'default': ''},
+
+    },
+    'actions': {
+        '.action-template.django.list': {},
+        '.action-template.django.retrieve': {},
+        '.action-template.django.create': {},
+        '.action-template.django.delete': {},
+        '.action-template.django.update': {},
+        '.action-template.django.partial_update': {},
+    },
+}
+
+ACCESS_POLICIES = {
+    '.collection.django': {},
+    'model': 'service.models.AccessPolicy',
+    'fields': {
+        'id': {
+            '.field.uuid': {},
+            '.flag.nowrite': {}},
+        'name': {
+            '.field.string': {},
+            '.flag.orderable': {},
+            '.flag.searchable': {}},
+        'access_mode': {
+            '.field.string': {},
+            '.flag.searchable': {},
+            '.flag.nullable.default': {}},
+        'payment_model': {
+            '.field.string': {},
+            '.flag.nullable.default': {}},
+       'pricing': {
+            '.field.string': {},
+            '.flag.nullable.default': {}},
+       'conditions': {
+            '.field.string': {},
+            '.flag.nullable.default': {}},
+       'geo_availability': {
+            '.field.string': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {},
+            '.flag.nullable.default': {}},
+       'access_policy_url': {
+            '.field.string': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {},
+            '.flag.nullable.default': {}},
+    },
+    'actions': {
+        '.action-template.django.list': {},
+        '.action-template.django.retrieve': {},
+        '.action-template.django.create': {},
+        '.action-template.django.delete': {},
+        '.action-template.django.update': {},
+        '.action-template.django.partial_update': {},
+    },
+}
+
+
+FEDERATION_MEMBERS = {
+    '.collection.django': {},
+    'model': 'service.models.FederationMember',
+    'fields': {
+        'id': {
+            '.field.uuid': {},
+            '.flag.nowrite': {}},
+        'name': {
+            '.field.string': {},
+            '.flag.orderable': {},
+            '.flag.searchable': {}},
+        'webpage': {
+            '.field.string': {},
+            '.flag.searchable': {},
+            '.flag.nullable.default': {}},
+        'logo': {
+            '.field.file': {},
+            'default': ''},
+        'country': {
+            '.field.string': {},
+            '.flag.orderable': {},
+            '.flag.nullable.default': {}},
+    },
+    'actions': {
+        '.action-template.django.list': {},
+        '.action-template.django.retrieve': {},
+        '.action-template.django.create': {},
+        '.action-template.django.delete': {},
+        '.action-template.django.update': {},
+        '.action-template.django.partial_update': {},
+    },
+}
+
 
 APP_CONFIG = {
     '.apimas_app': {},
@@ -1098,7 +1371,7 @@ APP_CONFIG = {
                 'external_services': EXTERNAL_SERVICES,
                 'user-roles': USER_ROLES,
                 'user_customers': USER_CUSTOMERS,
-                'service-areas': SERVICE_AREAS,
+                'service-categories': SERVICE_CATEGORIES,
                 'service-admins': SERVICE_ADMINS,
                 'custom-users': CUSTOM_USERS,
                 'contact-information': CONTACT_INFORMATION,
@@ -1109,6 +1382,9 @@ APP_CONFIG = {
                 'component-implementation-details': COMPONENT_IMPLEMENTATION_DETAILS,
                 'component-implementation-detail-links': COMPONENT_IMPLEMENTATION_DETAIL_LINKS,
                 'my-services': MY_SERVICES,
+                'providers': ORGANISATIONS,
+                'access-policies': ACCESS_POLICIES,
+                'federation-members': FEDERATION_MEMBERS,
             },
         },
     },