+++ tests/test_service-versions.py
@@ -1,4 +1,3 @@
-import json
 from agora.testing import *
 from service.models import ServiceStatus
 
@@ -37,14 +36,13 @@ def test_serviceversion_create_edit(serviceadmin, serviceadmin_id, superadmin):
     s2 = resp.json()['id']
 
     resp = superadmin.post(sa_url, {'service': s2, 'admin': serviceadmin_id})
-    status, created = ServiceStatus.objects.get_or_create(value='st',
-                                                          order="2")
+    status, created = ServiceStatus.objects.get_or_create(value='st')
 
     SV_DATA_1 = {
         'version': '1.1',
         'id_service': s1,
         'status': status.id,
-        'usage_policy_has': True,
+        'terms_of_use_has': True,
         'user_documentation_has': True,
         'privacy_policy_has': True,
         'operations_documentation_has': True,
@@ -61,7 +59,7 @@ def test_serviceversion_create_edit(serviceadmin, serviceadmin_id, superadmin):
         'version': '1.1',
         'id_service': s1,
         'status': status.id,
-        'usage_policy_has': True,
+        'terms_of_use_has': True,
         'user_documentation_has': True,
         'privacy_policy_has': True,
         'operations_documentation_has': True,
@@ -72,12 +70,11 @@ def test_serviceversion_create_edit(serviceadmin, serviceadmin_id, superadmin):
         'decommissioning_procedure_has': True,
     }
 
-
     SV_DATA_2 = {
         'version': '1.2',
         'id_service': s2,
         'status': status.id,
-        'usage_policy_has': True,
+        'terms_of_use_has': True,
         'user_documentation_has': True,
         'privacy_policy_has': True,
         'operations_documentation_has': True,
@@ -94,7 +91,7 @@ def test_serviceversion_create_edit(serviceadmin, serviceadmin_id, superadmin):
         'version': '1.2',
         'id_service': s2,
         'status': status.id,
-        'usage_policy_has': True,
+        'terms_of_use_has': True,
         'user_documentation_has': True,
         'privacy_policy_has': True,
         'operations_documentation_has': True,
@@ -104,8 +101,6 @@ def test_serviceversion_create_edit(serviceadmin, serviceadmin_id, superadmin):
         'disaster_recovery_plan_has': True,
         'decommissioning_procedure_has': True,
     }
-
-
 
     # CREATE
     resp = serviceadmin.post(sv_url, SV_DATA_1)
@@ -120,11 +115,10 @@ def test_serviceversion_create_edit(serviceadmin, serviceadmin_id, superadmin):
 
     resp = serviceadmin.post(sv_url, SV_DATA_2)
     assert resp.status_code == 403
-     
+
     resp = serviceadmin.post(sv_url, SV_DATA_2_LIMITED)
     assert resp.status_code == 201
     sv2_id = resp.json()['id']
-
 
     # DELETE
     resp = serviceadmin.delete(sv_url + sv1_id + '/')
@@ -137,7 +131,7 @@ def test_serviceversion_create_edit(serviceadmin, serviceadmin_id, superadmin):
     assert resp.status_code == 204
 
     # Clean up
-    superadmin.delete(service_url+s1+'/')
-    superadmin.delete(service_url+s2+'/')
-    superadmin.delete(sv_url+sv2_id+'/')
-    superadmin.delete(status_url+str(status.id)+'/')
+    superadmin.delete(service_url + s1 + '/')
+    superadmin.delete(service_url + s2 + '/')
+    superadmin.delete(sv_url + sv2_id + '/')
+    superadmin.delete(status_url + str(status.id) + '/')